// src/apps/interfaces/AppInterface.ts
export interface GeneralDataInterface {
    label: string;
    value: number | string;
    init?: () => Promise<void>;
    staticValue?: () => string | number;
    asyncValue?: () => Promise<string | number>;
}